Fixing type conversion issues in Smac Planner changes

+++ nav2_smac_planner/src/a_star.cpp
@@ -230,9 +230,12 @@ void AStarAlgorithm<NodeT>::setGoal(
   switch (goal_heading_mode) {
     case GoalHeadingMode::DEFAULT: {
         // add a single goal node with single heading
-        auto goal = addToGraph(NodeT::getIndex(mx, my, dim_3));
-        goal->setPose(typename NodeT::Coordinates(
-          static_cast<int>(mx), static_cast<int>(my), static_cast<int>(dim_3)));
+        auto goal = addToGraph(
+          NodeT::getIndex(
+            static_cast<unsigned int>(mx),
+            static_cast<unsigned int>(my),
+            dim_3));
+        goal->setPose(typename NodeT::Coordinates(mx, my, static_cast<float>(dim_3)));
         goals_state.push_back({goal, true});
         break;
       }
@@ -240,16 +243,22 @@ void AStarAlgorithm<NodeT>::setGoal(
     case GoalHeadingMode::BIDIRECTIONAL: {
         // Add two goals, one for each direction
         // add goal in original direction
-        auto goal = addToGraph(NodeT::getIndex(mx, my, dim_3));
-        goal->setPose(typename NodeT::Coordinates(
-          static_cast<int>(mx), static_cast<int>(my), static_cast<int>(dim_3)));
+        auto goal = addToGraph(
+          NodeT::getIndex(
+            static_cast<unsigned int>(mx),
+            static_cast<unsigned int>(my),
+            dim_3));
+        goal->setPose(typename NodeT::Coordinates(mx, my, static_cast<float>(dim_3)));
         goals_state.push_back({goal, true});
 
         // Add goal node in opposite (180°) direction
         unsigned int opposite_heading = (dim_3 + (num_bins / 2)) % num_bins;
-        auto opposite_goal = addToGraph(NodeT::getIndex(mx, my, opposite_heading));
-        opposite_goal->setPose(typename NodeT::Coordinates(
-          static_cast<int>(mx), static_cast<int>(my), static_cast<int>(opposite_heading)));
+        auto opposite_goal = addToGraph(NodeT::getIndex(
+            static_cast<unsigned int>(mx),
+            static_cast<unsigned int>(my),
+            opposite_heading));
+        opposite_goal->setPose(
+          typename NodeT::Coordinates(mx, my, static_cast<float>(opposite_heading)));
         goals_state.push_back({opposite_goal, true});
         break;
       }
@@ -260,9 +269,12 @@ void AStarAlgorithm<NodeT>::setGoal(
 
         // Add goal nodes for all headings
         for (unsigned int i = 0; i < num_bins; ++i) {
-          auto goal = addToGraph(NodeT::getIndex(mx, my, i));
-          goal->setPose(typename NodeT::Coordinates(
-            static_cast<float>(mx), static_cast<float>(my), static_cast<float>(i)));
+          auto goal = addToGraph(
+            NodeT::getIndex(
+              static_cast<unsigned int>(mx),
+              static_cast<unsigned int>(my),
+              i));
+          goal->setPose(typename NodeT::Coordinates(mx, my, static_cast<float>(i)));
           goals_state.push_back({goal, true});
         }
         break;
